Genesis
In the beginning Man created the computer. Now the computer was
formless and empty, darkness was over the surface of the deep, and the
Spirit of Man was hovering over the terminal.

And Man said, "Let there be lisp," and there was lisp. Man saw that
the lisp was good.