readme
The recommended way to become familiar with the academy is to start
with the first commit either on github or with local git tools and to
work your way through commits attempting to understand at least the
jist if not the entirety of what is going on. Or, alternatively, you
could just fire it up and with the help of HELP, explore.