pedantic-if
IFs with no else clause or a NIL else clause should always be replaced
with a WHEN.

Adhering to style has concrete advantages. The advantage is that it
decreases the mental effort required for a human to parse and
understand the code. In this case, if adhered to, every IF you see
will always have an else clause. You can count on it. Ignore this
style and every time you see an IF you wont know until you read
further in the code. A definite advantage and valuable.

Here we shadow the standard IF with our own that warns when we violate
these style rules.

> (defun foo (a) (if (eq a :foo) :baz))
; in: DEFUN FOO
;     (ACADEMY::IF (EQ ACADEMY::A :FOO) :BAZ)
;
; caught WARNING:
;   A single legged IF should be replaced with a WHEN.
;
; compilation unit finished
;   caught 1 WARNING condition
FOO

> (foo :foo)
:BAZ

> (foo 123)
NIL

> (if :foo :baz nil)
WARNING: An if's else of NIL should be replaced with a WHEN.

This commit demonstrates

  * Shadowing core CL functionality with our own cooked up versions.
  * Using the optional SUPPLIED-P-PARAMETER to distinguish between a
    supplied NIL and no argument.